Session mentorat - Création deleteaccount2

+++ backend/controllers/user.js
@@ -17,7 +17,7 @@ exports.signup = (req, res) => {
         .then(() => res.status(201).json({ message: "Félicitations ! Votre compte utilisateur a été créé." }))
         .catch(error => {
           console.error(error);
-          res.status(400).json({ message: "Cette adresse mail est déjà utilisée, veuillez en saisir une autre." })
+          res.status(400).json({ message: "Cette adresse mail est déjà utilisée." })
         })
     })
     .catch(error => {
@@ -66,30 +66,55 @@ exports.login = (req, res) => {
 };
 
 exports.updateProfile = (req, res) => {
+
   const newProfilePic = req.file;
+  console.log(newProfilePic)
   const userObject = req.file ?
   {
-    ...JSON.parse(req.body.user),
     ...JSON.parse(req.body),
     profileImageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
   } : { ...req.body };
-  console.log("userObject");
-  console.log(userObject);
+  console.log(userObject)
 
   sequelize.models.User.findOne({ where: { id: req.params.id } })
   .then(user => {
     if (newProfilePic) {
       const fileToBeDeleted = user.profileImageUrl.split('/images')[1];
-      user.update({ ...userObject }, { where: { id: req.params.id } })
-        .then(() => {
-          res.status(200).json({ message: "Profil mis à jour" })
-          fs.unlink(`images/${fileToBeDeleted}`, (err) => {
-              if (err) {
-                  console.error(err);
-              }
+
+      bcrypt.hash(req.body.password, 10)
+          .then(hash =>{
+            user.update({
+              ...userObject,
+              password: hash
+            }, { where: { id: req.params.id } })
+              .then(user => {
+                user.save()
+                .then(() => {
+                  res.status(200).json({ message: "Votre profil a été mis à jour et sauvegardé" })
+                  if (fileToBeDeleted) {
+                    fs.unlink(`images/${fileToBeDeleted}`, (err) => {
+                      if (err) {
+                          console.error(err);
+                      }
+                  })
+
+                  return;
+                  }
+                })
+                .catch(error => {
+                  console.error(error);
+                  res.status(500).json({ message: "Erreur SERVEUR" })
+                })
+              })
+              .catch(error => {
+                console.error(error);
+                res.status(500).json({ message: "Erreur SERVEUR" })
+              })
           })
-        })
-        .catch(error => res.status(400).json({ message: error }))
+          .catch(error => {
+            console.error(error);
+            res.status(500).json({ message: "Erreur SERVEUR" })
+          })
 
       return;
     } else {
@@ -101,8 +126,11 @@ exports.updateProfile = (req, res) => {
           }, { where: { id: req.params.id } })
             .then(user => {
               user.save()
-                .then(() => res.status(200).json({ message: "Profil mis à jour et sauvegardé" }))
-                .catch()
+                .then(() => res.status(200).json({ message: "Votre profil a été mis à jour et sauvegardé" }))
+                .catch(error => {
+                  console.error(error);
+                  res.status(500).json({ message: "Erreur SERVEUR" })
+                })
             })
             .catch(error => {
               console.error(error);
@@ -125,36 +153,60 @@ exports.updateProfile = (req, res) => {
 };
 
 exports.deleteAccount = (req, res) => {
-  console.log(req.auth);
-  sequelize.models.User.findOne({ id: req.params.id })
-    .then(user => {
-      const userId = user.id;
 
-      console.log("User ID");
-      console.log(userId);
-      // Vérifier que l'id du user voulant supprimer le compte est le même que l'id dans l'URL
-      if (!user) {
-        res.status(404).json({ message: "Utilisateur non reconnu" });
+  sequelize.models.User.findOne({ where: { 
+    id: req.body.id,
+    email: req.body.email
+  } })
+  .then(user => {
+    //console.log("user.id", user.id);
+    //console.log("req.auth.userId", user.id)
+    // Vérifier que l'id du user voulant supprimer le compte est le même que l'id dans l'URL
+    if (!user) {
+      res.status(404).json({ message: "Utilisateur non reconnu" });
 
-        return;
-        // Vérifier que l'id du token est le même que celui du user qui veut supprimer le compte
-      } /*else if (user.id != req.auth.userId) {
-          res.status(403).json({ message: 'Requête non autorisée !' });
+      return;
+    } else if (req.params.id !== req.auth) {
+      // Vérifier que le token d'authentification est le même que celui du user qui veut supprimer le compte
+      res.status(403).json({ message: "Requête non autorisée !" });
 
-          return;
-      }*/
+      return;
+    }
+    
+    sequelize.models.User.destroy({ where: { id: req.body.id } })
+      .then(() => res.status(200).json({ message: "Le profil a été supprimé" }))
+      .catch(error => {
+        console.error(error);
+        res.status(500).json({ message: "Erreur SERVEUR" })
+      })     
+  })
+  .catch(error => {
+    console.error(error);
+    res.status(500).json({ message: "Erreur SERVEUR" })
+  })
+    
+};
 
-      console.log(user)
+exports.deleteAccount2 = (req, res) => {
+  console.log("req.body.id", req.body.id);
+  console.log("req.params.id", req.params.id);
+  console.log("req.auth", req.auth);
+  console.log("Test1", req.body.id != req.params.id);
 
-      user.destroy({ id: req.params.id })
-        .then(() => res.status(200).json({ message: "Le profil a été supprimé" }))
-        .catch(error => {
-          console.error(error);
-          res.status(500).json({ message: "Erreur SERVEUR" })
-        })      
+
+  if (req.body.id !== req.params.id || req.body.id !== req.auth) {
+    res.status(403).json({message: "Forbidden"})
+  }
+
+  sequelize.models.User.destroy({ where: { id: req.params.id } })
+    .then(() => {
+      res.status(200).json({ message : "profil supprimé" })
     })
     .catch(error => {
-      console.error(error);
-      res.status(500).json({ message: "Erreur SERVEUR" })
+      console.log(error);
+      res.status(500).json({ message: "Server down" })
     })
+
+  
+
 };
